Rename API serivces to a more intuitive one
