+++ src/api/atickets.js
@@ -1,6 +1,5 @@
-// src/api/atickets.js
 import { authFetch } from './authRefresh.js';
-import { BASE_URL } from './auth.js';
+import { BASE_URL } from './authServices.js';
 
 /**
  * Helper: Parse JSON and throw errors uniformly
