+++ src/api/auser.js
@@ -1,4 +1,3 @@
-// src/api/auser.js
 import { authFetch } from './authRefresh.js';
 
 const API_BASE_URL = 'http://localhost:3000/api'
