+++ src/api/authServices.js
@@ -1,4 +1,3 @@
-// src/api/auth.js
 const API_BASE_URL = 'https://eventregistrationsystem-backend.onrender.com/api';
 export const BASE_URL = import.meta.env.VITE_API_BASE_URL || API_BASE_URL;
