+++ src/views/admin/Event/EventFormView.vue
@@ -4,7 +4,7 @@ import { useRoute, useRouter } from 'vue-router'
 import AdminLayout from '@/layouts/AdminLayout.vue'
 
 // Importing events and ticket related interfaces from the API layer (both internally using authFetch for authorisation transfer)
-import { fetchEventDetails, createEvent, updateEvent } from '@/api/aevents.js'
+import { fetchEventDetails, createEvent, updateEvent } from '@/api/eventServices.js'
 import { fetchTicketTypes, createTicketType, updateTicketType } from '@/api/atickets.js'
