+++ src/views/admin/Event/EventsList.vue
@@ -152,7 +152,7 @@
 import { ref, computed, onMounted } from 'vue'
 import { useRouter } from 'vue-router'
 import AdminLayout from '@/layouts/AdminLayout.vue'
-import { fetchEvents, deleteEvent as deleteEventApi } from '@/api/aevents.js'
+import { fetchEvents, deleteEvent as deleteEventApi } from '@/api/eventServices.js'
 
 // Router
 const router = useRouter()
