+++ src/views/admin/User/UserEdit.vue
@@ -3,7 +3,7 @@ import { ref, onMounted } from "vue";
 import { useRoute, useRouter } from "vue-router";
 import AdminLayout from "@/layouts/AdminLayout.vue";
 import { usersMockData } from "@/mock/usersMock.js";
-import { getUserById, updateUser } from "@/api/users";
+import { getUserById, updateUser } from "@/api/userServices";
 
 // 获取路由和路由器实例
 const route = useRoute();
